rewrite for RPMs and Yum
Rewrite RepoCollector to collect both Yum repos and Debian repos.

RepoCollector is a generalized repository collector now, without
externally-exposed attrs like "is_debian_repo".

Add DebRepo and RpmRepo helpers to discover the files we need to sign:

  - .repomd ("repomd.xml" files)
  - .rpms (".rpm" files)
  - .releases ("Release" files)

Centralize the general signing dispatcher logic to the Base backend
class, and update the Gpg and RpmSign backends with the
implementation-specific methods for each artifact.

Update the fixtures to be files in Git rather than dynamically
constructing repos with Python code and writing to a tmpdir. This makes
it easier to understand what an example repository layout looks like
(eg you can browse it with "ls -R".)